Refactoring and debugging of various parts of the dataloader and api
