+++ api/endpoints/utils.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Query
 from .endpoint_utils import execute_query
-from ..queries import main_queries
+from ..queries import utils_queries
 from ..utils import create_cleaned_limit_collection
 from ..search import search_utils
 from .models.shared import CountMatchesInput
@@ -22,7 +22,7 @@ async def get_counts_for_file(input: CountMatchesInput):
     )
 
     query_graph_result = execute_query(
-        main_queries.QUERY_COUNT_MATCHES,
+        utils_queries.QUERY_COUNT_MATCHES,
         bind_vars={
             "file_name": input.file_name,
             "score": input.score,
@@ -45,7 +45,7 @@ async def get_folios_for_file(
     suttas/PTS nrs/segments (PLI) / segments (SKT)
     """
     query_graph_result = execute_query(
-        main_queries.QUERY_FOLIOS,
+        utils_queries.QUERY_FOLIOS,
         bind_vars={"file_name": file_name},
     )
     folios = query_graph_result.result[0]
@@ -62,7 +62,7 @@ async def get_displayname_for_segmentnr(
     """
     filename = segmentnr.split(":")[0]
     query_graph_result = execute_query(
-        main_queries.QUERY_DISPLAYNAME,
+        utils_queries.QUERY_DISPLAYNAME,
         bind_vars={"filename": filename},
     )
     displayname = query_graph_result.result[0]
@@ -93,7 +93,7 @@ async def get_multilingual(
     """
     query_result = {"langList": []}
     query_displayname = execute_query(
-        main_queries.QUERY_MULTILINGUAL_LANGS,
+        utils_queries.QUERY_MULTILINGUAL_LANGS,
         bind_vars={"file_name": file_name},
         raw_results=True,
     )
