+++ dataloader/dataloader_constants.py
@@ -27,8 +27,12 @@ COLLECTION_LANGUAGES = "languages"
 COLLECTION_FILES = "files"
 COLLECTION_MENU_COLLECTIONS = "menu_collections"
 COLLECTION_MENU_CATEGORIES = "menu_categories"
-COLLECTION_FILES_PARALLEL_COUNT = "files_parallel_count"
-COLLECTION_CATEGORIES_PARALLEL_COUNT = "categories_parallel_count"
+
+
+GLOBAL_STATS_CATEGORIES = "global_stats_categories"
+GLOBAL_STATS_FILES = "global_stats_files"
+
+
 
 COLLECTION_SEARCH_INDEX_TIB = "search_index_tib"
 COLLECTION_SEARCH_INDEX_SKT = "search_index_skt"
@@ -87,9 +91,11 @@ COLLECTION_NAMES = (
     COLLECTION_FILES,
     COLLECTION_MENU_COLLECTIONS,
     COLLECTION_MENU_CATEGORIES,
-    COLLECTION_FILES_PARALLEL_COUNT,
-    COLLECTION_CATEGORIES_PARALLEL_COUNT,
+    GLOBAL_STATS_CATEGORIES,
+    GLOBAL_STATS_FILES,
 )
+
+
 
 # Edge collections
 EDGE_COLLECTION_LANGUAGE_HAS_COLLECTIONS = "language_has_collections"
