+++ dataloader/tasks.py
@@ -15,8 +15,6 @@ from dataloader_constants import (
     COLLECTION_NAMES,
     DEFAULT_SOURCE_URL,
     DEFAULT_TSV_URL,
-    EDGE_COLLECTION_NAMES,
-    GRAPH_COLLECTIONS_CATEGORIES,
     LANG_TIBETAN,
     LANG_PALI,
     LANG_CHINESE,
@@ -43,10 +41,11 @@ from load_parallels import (
     clean_parallels_for_language
 )
 
+from load_stats import load_global_stats_for_language
+
 from tasks_menu import (
     load_all_menu_collections,
     load_all_menu_categories,
-    create_collections_categories_graph,
 )
 
 from utils import get_database, get_system_database
@@ -54,7 +53,7 @@ from utils import get_database, get_system_database
 from clean_database import (
     clean_search_index_db,
     clean_all_collections_db,
-    clean_totals_collection_db,
+    clean_global_stats_db,
     clean_segment_collections_db,
     clean_menu_collections_db,
     clean_all_lang_db,
@@ -87,7 +86,7 @@ def create_db(c):
 
 @task(help={"collections": "Array of collections you'd like to create"})
 def create_collections(
-    c, collections=COLLECTION_NAMES, edge_collections=EDGE_COLLECTION_NAMES
+    c, collections=COLLECTION_NAMES
 ):
     """
     Create empty collections in database
@@ -102,11 +101,6 @@ def create_collections(
             db.create_collection(name)
         except CollectionCreateError as e:
             print(f"Error creating collection {name}: ", e)
-    for name in edge_collections:
-        try:
-            db.create_collection(name, edge=True)
-        except CollectionCreateError as e:
-            print("Error creating edge collection: ", e)
     print(f"created {collections} collections")
 
 
@@ -171,9 +165,9 @@ def load_parallels(c, root_url=DEFAULT_SOURCE_URL, lang=DEFAULT_LANGS, threaded=
     db = get_database()
     for clang in lang:
         print("LANG: ", clang)
-        load_parallels_for_language(
-            root_url, clang, db, thread_count if threaded else 1
-        )
+        #load_parallels_for_language(
+        #    root_url, clang, db, thread_count if threaded else 1
+        #)
         load_sorted_parallels_for_language(root_url, clang, db)
 
 @task
@@ -184,6 +178,22 @@ def clean_parallels(c, lang=DEFAULT_LANGS):
     for l in lang:
         clean_parallels_for_language(l, db)
         print("Parallel data cleaned for language ", l)
+
+
+@task
+def load_global_stats(c, root_url=DEFAULT_SOURCE_URL, lang=DEFAULT_LANGS):
+    db = get_database()
+    if lang != DEFAULT_LANGS:
+        lang = ["".join(lang)]
+    for l in lang:
+        print("Loading global stats for language: ", l)
+        load_global_stats_for_language(root_url, l, db)
+        print("Global stats loaded for language ", l)
+
+@task
+def clean_global_stats(c):
+    clean_global_stats_db()
+
 
 
 @task
@@ -241,14 +251,8 @@ def clean_pali(c):
         for name in COLLECTION_NAMES:
             current_name = name
             db.delete_collection(name)
-        for name in EDGE_COLLECTION_NAMES:
-            current_name = name
-            db.delete_collection(name)
-        db.delete_graph(GRAPH_COLLECTIONS_CATEGORIES)
     except CollectionDeleteError as e:
         print("Error deleting collection %s: " % current_name, e)
-    except GraphDeleteError as e:
-        print("couldn't remove graph. It probably doesn't exist.", e)
 
     print("all collections cleaned.")
 
@@ -329,7 +333,7 @@ def load_menu_files(c):
     db = get_database()
     load_all_menu_categories(db)
     load_all_menu_collections(db)
-    create_collections_categories_graph(db)
+    
 
     print("Menu data loading completed!")
