+++ dataloader/tasks_menu.py
@@ -10,65 +10,9 @@ from dataloader_validator import validate_json, METADATA_ROOT
 from dataloader_constants import (
     COLLECTION_MENU_COLLECTIONS,
     DEFAULT_LANGS,
-    COLLECTION_MENU_CATEGORIES,
-    EDGE_COLLECTION_COLLECTION_HAS_CATEGORIES,
-    GRAPH_COLLECTIONS_CATEGORIES,
+    COLLECTION_MENU_CATEGORIES,    
     COLLECTION_LANGUAGES,
-    EDGE_COLLECTION_LANGUAGE_HAS_COLLECTIONS,
-    EDGE_COLLECTION_CATEGORY_HAS_FILES,
-    COLLECTION_FILES,
 )
-
-
-def create_collections_categories_graph(db: StandardDatabase) -> None:
-    if db.has_graph(GRAPH_COLLECTIONS_CATEGORIES):
-        return
-
-    graph = db.create_graph(GRAPH_COLLECTIONS_CATEGORIES)
-    # Language -> Collections
-    graph.create_edge_definition(
-        edge_collection=EDGE_COLLECTION_LANGUAGE_HAS_COLLECTIONS,
-        from_vertex_collections=[COLLECTION_LANGUAGES],
-        to_vertex_collections=[COLLECTION_MENU_COLLECTIONS],
-    )
-    # Collection -> Categories
-    graph.create_edge_definition(
-        edge_collection=EDGE_COLLECTION_COLLECTION_HAS_CATEGORIES,
-        from_vertex_collections=[COLLECTION_MENU_COLLECTIONS],
-        to_vertex_collections=[COLLECTION_MENU_CATEGORIES],
-    )
-    # Category -> Files
-    graph.create_edge_definition(
-        edge_collection=EDGE_COLLECTION_CATEGORY_HAS_FILES,
-        from_vertex_collections=[COLLECTION_MENU_CATEGORIES],
-        to_vertex_collections=[COLLECTION_FILES],
-    )
-
-
-def create_edges_for_collection_has_categories(
-    language: str,
-    categories: dict,
-    collection_key: str,
-    edge_db_collection: EdgeCollection,
-) -> None:
-    for category_name in categories:
-        edge_db_collection.insert(
-            {
-                "_from": f"{COLLECTION_MENU_COLLECTIONS}/{collection_key}",
-                "_to": f"{COLLECTION_MENU_CATEGORIES}/{language}_{category_name}",
-            }
-        )
-
-
-def create_edges_for_language_has_collections(
-    language: str, collection_key: str, edge_db_collection: EdgeCollection
-):
-    edge_db_collection.insert(
-        {
-            "_from": f"{COLLECTION_LANGUAGES}/{language}",
-            "_to": f"{COLLECTION_MENU_COLLECTIONS}/{collection_key}",
-        }
-    )
 
 
 def load_menu_collection(
@@ -76,33 +20,20 @@ def load_menu_collection(
     language: str,
     collection_count: int,
     collections_db_collection: StandardCollection,
-    collection_has_categories_edge_db_collection: EdgeCollection,
-    language_has_collections_edge_db_collection: EdgeCollection,
 ):
     # Create document
     doc = {
         "_key": f"{language}_{menu_collection['collection']}",
         "language": language,
         "collectionnr": collection_count,
+        "categories": menu_collection["categories"],
+        "collection": menu_collection["collection"],
     }
-    # we won't need the categories array in the db, so let's store it away
-    categories = menu_collection["categories"]
-    del menu_collection["categories"]
-    doc.update(menu_collection)
-
+    
     try:
         # Create vertex documents..
         collections_db_collection.insert(doc)
-        # ..and edges
-        create_edges_for_collection_has_categories(
-            language,
-            categories,
-            doc["_key"],
-            collection_has_categories_edge_db_collection,
-        )        
-        create_edges_for_language_has_collections(
-            language, doc["_key"], language_has_collections_edge_db_collection
-        )
+
     except DocumentInsertError as e:
         print(doc["_key"])
         print("Could not load menu collection. Error: ", e)
@@ -112,12 +43,6 @@ def load_all_menu_collections(db: StandardDatabase):
     collections_db_collection = db.collection(COLLECTION_MENU_COLLECTIONS)
     languages_db_collection = db.collection(COLLECTION_LANGUAGES)
     # edge collections:
-    collection_has_categories_edge_db_collection = db.collection(
-        EDGE_COLLECTION_COLLECTION_HAS_CATEGORIES
-    )
-    language_has_collections_edge_db_collection = db.collection(
-        EDGE_COLLECTION_LANGUAGE_HAS_COLLECTIONS
-    )
 
     for language in DEFAULT_LANGS:
         try:
@@ -146,36 +71,21 @@ def load_all_menu_collections(db: StandardDatabase):
                     language,
                     collection_count,
                     collections_db_collection,
-                    collection_has_categories_edge_db_collection,
-                    language_has_collections_edge_db_collection,
                 )
                 collection_count += 1
             print("✓")
-
-
-def create_edges_for_category_has_files(
-    files: list, category_key: str, edge_db_collection: EdgeCollection
-):
-    for file_name in files:
-        edge_db_collection.insert(
-            {
-                "_from": f"{COLLECTION_MENU_CATEGORIES}/{category_key}",
-                "_to": f"{COLLECTION_FILES}/{file_name}",
-            }
-        )
 
 
 def load_menu_category(
     menu_category,
     category_count,
     language,
-    categories_db_collection,
-    category_has_files_edge_db_collection,
+    categories_db_collection
 ):
     doc = {
         "_key": f'{language}_{menu_category["category"]}',
         "language": language,
-        "categorynr": category_count,
+        "categorynr": category_count,        
     }
     files = menu_category["files"]
     del menu_category["files"]
@@ -183,9 +93,7 @@ def load_menu_category(
     try:
         categories_db_collection.insert(doc)
         categories_db_collection.add_hash_index(["category"], unique=False)
-        create_edges_for_category_has_files(
-            files, doc["_key"], category_has_files_edge_db_collection
-        )
+
     except DocumentInsertError as e:
         print(f"Could not load menu category {menu_category['category']}. Error: ", e)
     except IndexCreateError as e:
@@ -194,9 +102,6 @@ def load_menu_category(
 
 def load_all_menu_categories(db: StandardDatabase):
     categories_db_collection = db.collection(COLLECTION_MENU_CATEGORIES)
-    category_has_files_edge_db_collection = db.collection(
-        EDGE_COLLECTION_CATEGORY_HAS_FILES
-    )
     for language in DEFAULT_LANGS:
         categories_filepath = f"{METADATA_ROOT}{language}-categories.json"   # TODO: no hardcoding
         categories_schema = f"{METADATA_ROOT}/schemas/categories.json"   # TODO: no hardcoding
@@ -212,7 +117,6 @@ def load_all_menu_categories(db: StandardDatabase):
                     category_count,
                     language,
                     categories_db_collection,
-                    category_has_files_edge_db_collection,
                 )
                 category_count += 1
             print("✓")
